requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.4.1+cu118
accelerate==1.0.1
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.6.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
diffusers==0.30.3
fastapi==0.115.2
ffmpy==0.4.0
filelock==3.16.1
fsspec==2024.9.0
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
gradio==5.0.2
gradio_client==1.4.0
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.1
numpy==2.1.2
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.2
rsa==4.9
ruff==0.6.9
safetensors==0.4.5
semantic-version==2.10.0
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.39.2
sympy==1.13.3
tokenizers==0.20.1
tomlkit==0.12.0
tqdm==4.66.5
transformers==4.45.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.31.1
websockets==12.0
zipp==3.20.2 
google-cloud-vision==3.4.4
opencv-python==4.8.0.74
